requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==4.4.2
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
proglog==0.1.10
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
six==1.16.0
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
